Replaced r['comments'][index] with hard-coded 0
We encountered an issue when creating multiple records of the same name with different data.  The code block which executes to create multiple RecordEntry objects was attempting to use the incremented index value despite the fact that there is only ever a single comment associated with the records.  Hard-coding this to the first element corrected the failure.